+++ R/join.R
@@ -8,6 +8,18 @@
 #' perform a natural join, using all variables in common across `x` and `y`. A
 #' message lists the variables so that you can check they're correct; suppress
 #' the message by supplying `by` explicitly.
+#'
+#' `by` can take a character vector, like `c("x", "y")` if `x` and `y` are
+#' in both datasets. To join on variables that don't have the same name, use
+#' equalities in the character vector, like `c("x1" = "x2", "y")`. If you use
+#' a character vector, the join can only be done using strict equality.
+#'
+#' Finally, `by` can be a specification created by `dplyr::join_by()`. Contrary
+#' to the input as character vector shown above, `join_by()` uses unquoted column
+#' names, e.g `join_by(x1 == x2, y)`. It also uses equality and inequality
+#' operators `==`, `>` and similar. **For now, only equality operators are
+#' supported**.
+#'
 #' @param suffix If there are non-joined duplicate variables in `x` and `y`,
 #' these suffixes will be added to the output to disambiguate them. Should be a
 #' character vector of length 2.
@@ -18,13 +30,13 @@
 #' @examplesIf require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)
 #' test <- polars::pl$DataFrame(
 #'   x = c(1, 2, 3),
-#'   y = c(1, 2, 3),
+#'   y1 = c(1, 2, 3),
 #'   z = c(1, 2, 3)
 #' )
 #'
 #' test2 <- polars::pl$DataFrame(
 #'   x = c(1, 2, 4),
-#'   y = c(1, 2, 4),
+#'   y2 = c(1, 2, 4),
 #'   z2 = c(4, 5, 7)
 #' )
 #'
@@ -32,20 +44,23 @@
 #'
 #' test2
 #'
+#' # default is to use common columns, here "x" only
 #' left_join(test, test2)
 #'
-#' inner_join(test, test2)
+#' # we can specify the columns on which to join with join_by()...
+#' left_join(test, test2, by = join_by(x, y1 == y2))
 #'
-#' full_join(test, test2)
+#' # ... or with a character vector
+#' left_join(test, test2, by = c("x", "y1" = "y2"))
 #'
-#' # Show how the arg 'suffix' works:
+#' # we can customize the suffix of common column names not used to join
 #' test2 <- polars::pl$DataFrame(
 #'   x = c(1, 2, 4),
-#'   y = c(1, 2, 4),
+#'   y1 = c(1, 2, 4),
 #'   z = c(4, 5, 7)
 #' )
 #'
-#' left_join(test, test2, by = c("x", "y"), suffix = c("_left", "_right"))
+#' left_join(test, test2, by = "x", suffix = c("_left", "_right"))
 
 left_join.DataFrame <- function(x, y, by = NULL, copy = NULL,
                                 suffix = c(".x", ".y"), ..., keep = NULL) {
@@ -197,6 +212,23 @@ join_ <- function(x, y, by = NULL, how, suffix) {
     )
   }
 
+  if (inherits(by, "dplyr_join_by")) {
+    by <- unpack_join_by(by)
+  }
+
+  if (!is.null(names(by))) {
+    for (i in seq_along(by)) {
+      if (names(by)[i] == "") {
+        names(by)[i] <- by[i]
+      }
+    }
+    left_on <- names(by)
+    right_on <- unname(by)
+  } else {
+    left_on <- by
+    right_on <- by
+  }
+
   dupes <- intersect(
     setdiff(pl_colnames(x), by),
     setdiff(pl_colnames(y), by)
@@ -207,9 +239,9 @@ join_ <- function(x, y, by = NULL, how, suffix) {
     (inherits(x, "LazyFrame") && inherits(y, "LazyFrame"))
   ) {
     if (how == "right") {
-      out <- y$join(other = x, on = by, how = "left")
+      out <- y$join(other = x, left_on = left_on, right_on = right_on, how = "left")
     } else {
-      out <- x$join(other = y, on = by, how = how)
+      out <- x$join(other = y, left_on = left_on, right_on = right_on, how = how)
     }
   }
 
@@ -220,4 +252,17 @@ join_ <- function(x, y, by = NULL, how, suffix) {
   } else {
     out
   }
+}
+
+
+unpack_join_by <- function(by) {
+  if (!all(by$condition == "==")) {
+    rlang::abort(
+      "`tidypolars` doesn't support inequality conditions in `join_by()` yet.",
+      call = rlang::caller_env(2)
+    )
+  }
+  out <- by$y
+  names(out) <- by$x
+  out
 }